+++ src/app.tsx
@@ -3,12 +3,12 @@ import { BrowserRouter } from "react-router-dom";
 
 import { Loading } from "~/components/loading/loading";
 import { AppRoutes } from "~/routes/routes";
-import { initialServicesState } from "~/state";
+import { initialGlobalProps } from "~/state";
 import { GlobalProvider } from "~/state/global/provider";
 
 export const App = (): ReactElement => (
   <BrowserRouter>
-    <GlobalProvider value={initialServicesState}>
+    <GlobalProvider value={initialGlobalProps}>
       <Loading />
       <AppRoutes />
     </GlobalProvider>