+++ src/data/services/user/user-service.ts
@@ -1,0 +1,5 @@
+import { UserListeners } from "~/services/user/types";
+
+export interface UserService {
+  getUserById: (id: string, listeners: UserListeners) => Promise<void>;
+}