+++ src/state/global/provider/types.ts
@@ -1,0 +1,8 @@
+import { StyledGuideProps } from "react-styled-guide";
+
+import { ServicesTypes } from "~/state/types";
+
+export type GlobalProviderProps = {
+  services: ServicesTypes;
+  styleGuide: StyledGuideProps;
+};