+++ src/state/theme/hook/types.ts
@@ -1,8 +1,0 @@
-import { ThemeDispatch, ThemeState } from "~/state/theme";
-
-type ThemeActionTypes = {
-  setDarkTheme: () => void;
-  setLightTheme: () => void;
-};
-
-export type UseThemeTypes = ThemeState & ThemeActionTypes & { dispatch: ThemeDispatch };