+++ src/style-guide/styled.d.ts
@@ -1,6 +1,0 @@
-import { StyleGuide } from "~/style-guide";
-
-declare module "styled-components" {
-  // eslint-disable-next-line @typescript-eslint/no-empty-interface
-  export interface DefaultTheme extends StyleGuide {}
-}