+++ tests/pages/base/base.test.tsx
@@ -3,14 +3,14 @@ import { BrowserRouter } from "react-router-dom";
 import { ThemeProvider } from "styled-components";
 
 import { Base } from "~/pages/base/base";
-import { defaultStyleGuide } from "~/style-guide";
+import initialStyleGuide from "~/style-guide";
 
 describe("<Base />", () => {
   it("should be able to mount <Base /> component", async () => {
     // given
     const { asFragment } = render(
       <BrowserRouter>
-        <ThemeProvider theme={defaultStyleGuide}>
+        <ThemeProvider theme={initialStyleGuide}>
           <Base />
         </ThemeProvider>
       </BrowserRouter>